refactor(backend): isolate Tooling lifecycle policy
