+++ src-tauri/src/services/tooling.rs
@@ -5,7 +5,22 @@ use crate::services::ProviderService;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
 
+mod lifecycle;
 mod versions;
+
+#[cfg(test)]
+use lifecycle::*;
+use lifecycle::{
+    build_tool_lifecycle_command, chain_update_commands, is_lifecycle_writable,
+    normalize_requested_tools, npm_install_command_for, official_update_args,
+    tool_action_shell_command, LifecycleCommandShell, ToolLifecycleAction,
+};
+#[cfg(target_os = "windows")]
+use lifecycle::{grok_install_windows_command, win_double_quote, windows_cmd_double_quote_arg};
+#[cfg(target_os = "macos")]
+use lifecycle::{
+    CLAUDE_INSTALL_UNIX, GROK_INSTALL_UNIX, HERMES_INSTALL_UNIX, OPENCODE_INSTALL_UNIX,
+};
 
 use versions::{
     elevated_windows_tool_version_unavailable, extract_version, get_single_tool_version_impl,
@@ -69,12 +84,6 @@ fn detected_tool_execution_boundary_for(formal_windows_build: bool) -> Result<()
     } else {
         Ok(())
     }
-}
-
-/// 生命周期写权限比只读探测更窄。Codex 保留在 `VALID_TOOLS`，因此版本和安装分布
-/// 诊断仍可用；但不得规划或执行 install/update/repair 命令。
-fn is_lifecycle_writable(tool: &str) -> bool {
-    tool != "codex"
 }
 
 pub async fn get_tool_versions(tools: Option<Vec<String>>) -> Result<Vec<ToolVersion>, String> {
@@ -309,336 +318,6 @@ fn decode_windows_command_output(bytes: &[u8]) -> String {
     }
 
     String::from_utf8_lossy(bytes).into_owned()
-}
-
-fn normalize_requested_tools(tools: &[String]) -> Vec<&'static str> {
-    let set: std::collections::HashSet<&str> = tools.iter().map(|s| s.as_str()).collect();
-    VALID_TOOLS
-        .iter()
-        .copied()
-        .filter(|tool| set.contains(tool))
-        .collect()
-}
-
-/// Q4 提权 CLI 生命周期动作白名单。此枚举是 install/update 的完整可写集合；
-/// 它不代表一般外部应用、终端或编辑器启动能力。
-#[derive(Debug, Clone, Copy)]
-enum ToolLifecycleAction {
-    Install,
-    Update,
-}
-
-impl FromStr for ToolLifecycleAction {
-    type Err = String;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value {
-            "install" => Ok(Self::Install),
-            "update" => Ok(Self::Update),
-            _ => Err(format!("Unsupported tool action: {value}")),
-        }
-    }
-}
-
-/// 私有、确定性的白名单命令构造器。前端不能传入命令行，且工具集合会再次由
-/// `VALID_TOOLS` / `is_lifecycle_writable` 收敛后才抵达执行边界。
-fn build_tool_lifecycle_command(
-    tools: &[&str],
-    action: ToolLifecycleAction,
-) -> Result<String, String> {
-    let mut lines = Vec::new();
-
-    #[cfg(target_os = "macos")]
-    {
-        // set -e 让任一步失败即中止;set -o pipefail 保留为管道命令的兜底防线。
-        // 当前官方 installer 路径已避免 `curl | bash`,但未来若新增管道命令,
-        // 仍应让管道前段失败参与整条脚本判定。
-        lines.push("set -e".to_string());
-        lines.push("set -o pipefail".to_string());
-    }
-
-    #[cfg(target_os = "windows")]
-    lines.push("@echo off".to_string());
-
-    for tool in tools {
-        if !is_lifecycle_writable(tool) {
-            return Err(CODEX_CLI_LIFECYCLE_DISABLED_MESSAGE.to_string());
-        }
-
-        let label = tool_display_name(tool);
-        lines.push(format!("echo ========== {label} =========="));
-
-        let line = build_tool_action_line(tool, action)?;
-        lines.push(line);
-
-        #[cfg(target_os = "windows")]
-        lines.push("if errorlevel 1 exit /b %errorlevel%".to_string());
-
-        #[cfg(target_os = "macos")]
-        lines.push(String::new());
-    }
-
-    Ok(lines.join(if cfg!(target_os = "windows") {
-        "\r\n"
-    } else {
-        "\n"
-    }))
-}
-
-fn tool_display_name(tool: &str) -> &'static str {
-    match tool {
-        "claude" => "Claude Code",
-        "codex" => "Codex",
-        "gemini" => "Gemini CLI",
-        "grok" => "Grok Build",
-        "opencode" => "OpenCode",
-        "openclaw" => "OpenClaw",
-        "hermes" => "Hermes",
-        _ => "Unknown",
-    }
-}
-
-/// 官方 shell installer 都不用 `curl | bash` 这种 pipe 形式（仍然用 curl 下载，
-/// 只是先落到临时文件再交给 bash 执行）。这样不依赖调用方 shell 的 pipefail
-/// 语义，且能让 curl 失败稳定变成整条命令失败。
-#[cfg(target_os = "macos")]
-const CLAUDE_INSTALL_UNIX: &str =
-    "bash -c 'tmp=$(mktemp) && curl -fsSL https://claude.ai/install.sh -o $tmp && bash $tmp; status=$?; rm -f $tmp; exit $status'";
-#[cfg(target_os = "macos")]
-const OPENCODE_INSTALL_UNIX: &str =
-    "bash -c 'tmp=$(mktemp) && curl -fsSL https://opencode.ai/install -o $tmp && bash $tmp; status=$?; rm -f $tmp; exit $status'";
-#[cfg(target_os = "macos")]
-const GROK_INSTALL_UNIX: &str =
-    "bash -c 'tmp=$(mktemp) && curl -fsSL https://x.ai/cli/install.sh -o $tmp && bash $tmp; status=$?; rm -f $tmp; exit $status'";
-
-/// Hermes 官方安装器会自带/选择合适的 Python 运行时。不要再用
-/// `python3 -m pip ... || python -m pip ...`:Hermes PyPI 包要求 Python >=3.11,
-/// 但 macOS 系统 `python3` 常是 3.9,而 pyenv 下 `python` shim 还可能不存在,会把
-/// 真正的 Python 版本问题盖成 "python command exists in these Python versions"。
-const HERMES_INSTALL_UNIX: &str =
-    "bash -c 'tmp=$(mktemp) && curl -fsSL https://raw.githubusercontent.com/NousResearch/hermes-agent/main/scripts/install.sh -o $tmp && bash $tmp; status=$?; rm -f $tmp; exit $status'";
-const HERMES_UPDATE_UNIX: &str =
-    "hermes update || bash -c 'tmp=$(mktemp) && curl -fsSL https://raw.githubusercontent.com/NousResearch/hermes-agent/main/scripts/install.sh -o $tmp && bash $tmp; status=$?; rm -f $tmp; exit $status'";
-
-#[cfg(target_os = "windows")]
-const HERMES_INSTALL_WINDOWS_SCRIPT: &str =
-    "irm https://raw.githubusercontent.com/NousResearch/hermes-agent/main/scripts/install.ps1 | iex";
-#[cfg(target_os = "windows")]
-const GROK_INSTALL_WINDOWS_SCRIPT: &str = "irm https://x.ai/cli/install.ps1 | iex";
-
-#[cfg(target_os = "windows")]
-fn powershell_encoded_command(script: &str) -> String {
-    use base64::{engine::general_purpose::STANDARD, Engine as _};
-
-    let mut bytes = Vec::with_capacity(script.len() * 2);
-    for unit in script.encode_utf16() {
-        bytes.extend_from_slice(&unit.to_le_bytes());
-    }
-    STANDARD.encode(bytes)
-}
-
-#[cfg(target_os = "windows")]
-fn hermes_install_windows_command() -> String {
-    format!(
-        "powershell -NoProfile -ExecutionPolicy Bypass -EncodedCommand {}",
-        powershell_encoded_command(HERMES_INSTALL_WINDOWS_SCRIPT)
-    )
-}
-
-#[cfg(target_os = "windows")]
-fn grok_install_windows_command() -> String {
-    format!(
-        "powershell -NoProfile -ExecutionPolicy Bypass -EncodedCommand {}",
-        powershell_encoded_command(GROK_INSTALL_WINDOWS_SCRIPT)
-    )
-}
-
-#[cfg(target_os = "windows")]
-fn hermes_update_windows_command() -> String {
-    // fallback 是 powershell.exe，不是 .cmd/.bat；这里不需要 `call`。PowerShell 的
-    // `irm | iex` 已被 EncodedCommand 收进单一参数,避免 `cmd.exe` 解析管道符。
-    format!("hermes update || {}", hermes_install_windows_command())
-}
-
-#[derive(Debug, Clone, Copy)]
-enum LifecycleCommandShell {
-    #[cfg_attr(target_os = "windows", allow(dead_code))]
-    Posix,
-    #[cfg_attr(target_os = "macos", allow(dead_code))]
-    WindowsBatch,
-}
-
-fn npm_install_command_for(tool: &str) -> Option<&'static str> {
-    if !is_lifecycle_writable(tool) {
-        return None;
-    }
-
-    match tool {
-        "claude" => Some("npm i -g @anthropic-ai/claude-code@latest"),
-        "gemini" => Some("npm i -g @google/gemini-cli@latest"),
-        "grok" => Some("npm i -g @xai-official/grok@latest"),
-        "opencode" => Some("npm i -g opencode-ai@latest"),
-        "openclaw" => Some("npm i -g openclaw@latest"),
-        _ => None,
-    }
-}
-
-fn official_update_args(tool: &str) -> Option<&'static str> {
-    match tool {
-        "claude" | "grok" | "hermes" => Some("update"),
-        "openclaw" => Some("update --yes"),
-        "opencode" => Some("upgrade"),
-        _ => None,
-    }
-}
-
-fn bare_official_update_command(tool: &str) -> Option<String> {
-    official_update_args(tool).map(|args| format!("{tool} {args}"))
-}
-
-fn chain_update_commands(
-    primary: String,
-    fallback: String,
-    shell: LifecycleCommandShell,
-) -> String {
-    if fallback.trim().is_empty() {
-        return primary;
-    }
-    match shell {
-        LifecycleCommandShell::Posix => format!("{primary} || {fallback}"),
-        // 这段最终会被外层再包成 `call <command>`。fallback 若是 npm.cmd/pnpm.cmd,
-        // `||` 右侧也必须显式 `call`,否则批处理会转移控制权并跳过后续工具。
-        LifecycleCommandShell::WindowsBatch => format!("{primary} || call {fallback}"),
-    }
-}
-
-fn tool_action_shell_command_for_shell(
-    tool: &str,
-    action: ToolLifecycleAction,
-    shell: LifecycleCommandShell,
-) -> Option<String> {
-    if !is_lifecycle_writable(tool) {
-        return None;
-    }
-
-    // xAI's primary Windows distribution is the native PowerShell installer.
-    // Keep npm as the network/policy fallback, matching the POSIX installer chain.
-    #[cfg(target_os = "windows")]
-    if tool == "grok"
-        && matches!(action, ToolLifecycleAction::Install)
-        && matches!(shell, LifecycleCommandShell::WindowsBatch)
-    {
-        return Some(chain_update_commands(
-            grok_install_windows_command(),
-            npm_install_command_for(tool)?.to_string(),
-            shell,
-        ));
-    }
-
-    if tool == "hermes" {
-        return Some(
-            match (action, shell) {
-                (ToolLifecycleAction::Install, LifecycleCommandShell::Posix) => HERMES_INSTALL_UNIX,
-                (ToolLifecycleAction::Update, LifecycleCommandShell::Posix) => HERMES_UPDATE_UNIX,
-                #[cfg(target_os = "windows")]
-                (ToolLifecycleAction::Install, LifecycleCommandShell::WindowsBatch) => {
-                    return Some(hermes_install_windows_command());
-                }
-                #[cfg(target_os = "windows")]
-                (ToolLifecycleAction::Update, LifecycleCommandShell::WindowsBatch) => {
-                    return Some(hermes_update_windows_command());
-                }
-                #[cfg(target_os = "macos")]
-                (_, LifecycleCommandShell::WindowsBatch) => return None,
-            }
-            .to_string(),
-        );
-    }
-
-    let install = npm_install_command_for(tool)?;
-    match action {
-        ToolLifecycleAction::Install => Some(install.to_string()),
-        ToolLifecycleAction::Update => match prefers_official_update(tool, shell)
-            .then(|| bare_official_update_command(tool))
-            .flatten()
-        {
-            Some(update) => Some(chain_update_commands(update, install.to_string(), shell)),
-            None => Some(install.to_string()),
-        },
-    }
-}
-
-fn tool_action_shell_command(tool: &str, action: ToolLifecycleAction) -> Option<String> {
-    #[cfg(target_os = "windows")]
-    let shell = LifecycleCommandShell::WindowsBatch;
-    #[cfg(target_os = "macos")]
-    let shell = LifecycleCommandShell::Posix;
-
-    tool_action_shell_command_for_shell(tool, action, shell)
-}
-
-fn build_tool_action_line(tool: &str, action: ToolLifecycleAction) -> Result<String, String> {
-    #[cfg(target_os = "windows")]
-    {
-        // Windows 原生 update 锚定;install 走静态(install.sh 是 bash 脚本,Windows
-        //    无意义)。**`enumerate_tool_installations` 在这里 per-tool 重做、与前端
-        //    probe 阶段算过的结果不共享是 by design**:run_tool_lifecycle_action 是
-        //    独立 IPC 调用,不信任前端回传的命令字符串(避免命令注入面扩大);前端是
-        //    逐工具触发 lifecycle,batch 化会破坏"逐工具独立成败"的 UX。
-        let command = match action {
-            ToolLifecycleAction::Update => {
-                let installs = enumerate_tool_installations(tool);
-                installs_anchored_command(tool, &installs)
-                    .unwrap_or_else(|| static_fallback_command(tool))
-            }
-            ToolLifecycleAction::Install => {
-                static_fallback_command_for(tool, ToolLifecycleAction::Install)
-            }
-        };
-        if command.is_empty() {
-            return Err(format!("Unsupported tool action target: {tool}"));
-        }
-        // .bat 调用 .cmd/.bat 必须用 `call` 否则当前脚本被替换、后续 `if errorlevel`
-        // 行被跳过;对 .exe 加 call 无害(等同直接调用)。锚定命令头部可能是 .cmd
-        // (npm/pnpm)或 .exe(volta),静态命令头部是 `npm`(也是 .cmd)、`py` 等——
-        // 全部加 `call ` 前缀,风格统一且语义正确。含空格的头部已被 `win_quote_path_for_batch`
-        // 加上双引号,call 对带引号的路径解析正常。
-        Ok(format!("call {command}"))
-    }
-
-    #[cfg(target_os = "macos")]
-    {
-        // update 锚定到命令行实际命中的那处（写回同一个 node / brew / 原生安装器），
-        // 而非裸 `npm` 落到 PATH 第一个 npm；install 走「上游推荐 || npm 兜底」短路链
-        // （有 native installer 的工具如 claude/opencode/hermes），其余仍裸 npm。
-        let command = match action {
-            ToolLifecycleAction::Update => {
-                let installs = enumerate_tool_installations(tool);
-                installs_anchored_command(tool, &installs)
-                    .unwrap_or_else(|| static_fallback_command(tool))
-            }
-            ToolLifecycleAction::Install => install_command_for(tool),
-        };
-        if command.is_empty() {
-            return Err(format!("Unsupported tool action target: {tool}"));
-        }
-        Ok(command)
-    }
-}
-
-/// Windows 双引号包裹基础原语:无条件加引号 + 内部 `"` 转义为 `\"`。
-/// `windows_cmd_double_quote_arg`(给 cmd.exe 传单个参数用)与
-/// `win_quote_path_for_batch`(给锚定路径用)都基于它,避免两份 quoter 各自演化、
-/// 未来对同一路径产生不一致引用形态。镜像 POSIX 侧 `shell_single_quote` 与
-/// `quote_path_if_spaced` 的"重量基础 + 轻量条件包装"两层结构。
-#[cfg(target_os = "windows")]
-fn win_double_quote(value: &str) -> String {
-    format!("\"{}\"", value.replace('"', "\\\""))
-}
-
-#[cfg(target_os = "windows")]
-fn windows_cmd_double_quote_arg(value: &str) -> String {
-    win_double_quote(value)
 }
 
 /// 工具未安装时的统一错误文案。
